In order to allow for observing the matching process I have extracted the matching behaviour from RequestPattern to a new class RequestPatternMatcher and  introduced a MatcherObserver.
