+++ build.gradle
@@ -7,7 +7,7 @@ apply plugin: 'project-report'
 
 sourceCompatibility = 1.6
 group = 'com.github.tomakehurst'
-version = 1.57
+version = 1.58
 
 def versions = [
         jackson: '2.6.0',
