+++ src/main/java/com/github/tomakehurst/wiremock/WireMockServer.java
@@ -84,7 +84,8 @@ public class WireMockServer implements Container, Stubbing, Admin {
                 options.maxRequestJournalEntries(),
                 options.extensionsOfType(ResponseTransformer.class),
                 fileSource,
-                this
+                this,
+                options.requestPatternMatcher()
         );
 
         AdminRequestHandler adminRequestHandler = new AdminRequestHandler(
