+++ src/main/java/com/github/tomakehurst/wiremock/matching/MatcherObserver.java
@@ -1,0 +1,7 @@
+package com.github.tomakehurst.wiremock.matching;
+
+import com.github.tomakehurst.wiremock.http.Request;
+
+public interface MatcherObserver {
+    void onMatchingResult(boolean requestMatchesPattern, boolean urlMatches, boolean methodMatches, boolean requiredAbsentHeadersAreNotPresent, boolean headersMatch, boolean queryParametersMatch, boolean bodyMatches, Request request, RequestPattern requestPattern);
+}
