+++ src/main/java/com/github/tomakehurst/wiremock/matching/RequestPattern.java
@@ -18,25 +18,13 @@ package com.github.tomakehurst.wiremock.matching;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize.Inclusion;
 import com.github.tomakehurst.wiremock.common.Json;
-import com.github.tomakehurst.wiremock.http.HttpHeader;
-import com.github.tomakehurst.wiremock.http.QueryParameter;
-import com.github.tomakehurst.wiremock.http.Request;
 import com.github.tomakehurst.wiremock.http.RequestMethod;
-import com.google.common.base.Function;
-import com.google.common.base.Optional;
-import com.google.common.base.Predicate;
-import com.google.common.collect.ImmutableSet;
 
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
-import static com.github.tomakehurst.wiremock.common.LocalNotifier.notifier;
-import static com.github.tomakehurst.wiremock.http.RequestMethod.ANY;
-import static com.github.tomakehurst.wiremock.matching.ValuePattern.matching;
 import static com.google.common.base.Predicates.notNull;
 import static com.google.common.collect.FluentIterable.from;
-import static com.google.common.collect.Iterables.*;
 import static com.google.common.collect.Maps.newLinkedHashMap;
 import static java.util.Arrays.asList;
 
@@ -91,92 +79,6 @@ public class RequestPattern {
         if (from(asList(url, urlPath, urlPattern, urlPathPattern)).filter(notNull()).size() > 1) {
 			throw new IllegalStateException("Only one of url, urlPattern, urlPath or urlPathPattern may be set");
 		}
-	}
-
-	public boolean isMatchedBy(Request request) {
-		return (urlIsMatch(request) &&
-				methodMatches(request) &&
-                requiredAbsentHeadersAreNotPresentIn(request) &&
-				headersMatch(request) &&
-                queryParametersMatch(request) &&
-				bodyMatches(request));
-	}
-
-    private boolean urlIsMatch(Request request) {
-		String candidateUrl = request.getUrl();
-		boolean matched;
-		if (url != null) {
-            matched = url.equals(candidateUrl);
-        } else if (urlPattern != null) {
-			matched = candidateUrl.matches(urlPattern);
-		} else if (urlPathPattern != null) {
-			matched = candidateUrl.matches(urlPathPattern.concat(".*"));
-		} else {
-            matched = candidateUrl.startsWith(urlPath);
-        }
-
-		return matched;
-	}
-
-	private boolean methodMatches(Request request) {
-		boolean matched = method.equals(ANY) || request.getMethod().equals(method);
-		if (!matched) {
-			notifier().info(String.format("URL %s is match, but method %s is not", request.getUrl(), request.getMethod()));
-		}
-
-		return matched;
-	}
-
-    private boolean requiredAbsentHeadersAreNotPresentIn(final Request request) {
-        return !any(requiredAbsentHeaderKeys(), new Predicate<String>() {
-            public boolean apply(String key) {
-                return request.getAllHeaderKeys().contains(key);
-            }
-        });
-    }
-
-    private Set<String> requiredAbsentHeaderKeys() {
-        if (headerPatterns == null) {
-            return ImmutableSet.of();
-        }
-
-        return ImmutableSet.copyOf(filter(transform(headerPatterns.entrySet(), TO_KEYS_WHERE_VALUE_ABSENT), REMOVING_NULL));
-    }
-
-	private boolean headersMatch(final Request request) {
-        return noHeadersAreRequiredToBePresent() ||
-                all(headerPatterns.entrySet(), matchHeadersIn(request));
-	}
-
-    private boolean queryParametersMatch(Request request) {
-        return (queryParamPatterns == null ||
-                all(queryParamPatterns.entrySet(), matchQueryParametersIn(request)));
-    }
-
-    private boolean noHeadersAreRequiredToBePresent() {
-        return headerPatterns == null || allHeaderPatternsSpecifyAbsent();
-    }
-
-    private boolean allHeaderPatternsSpecifyAbsent() {
-        return size(filter(headerPatterns.values(), new Predicate<ValuePattern>() {
-            public boolean apply(ValuePattern headerPattern) {
-                return !headerPattern.nullSafeIsAbsent();
-            }
-        })) == 0;
-    }
-
-    private boolean bodyMatches(Request request) {
-		if (bodyPatterns == null) {
-			return true;
-		}
-
-		boolean matches = all(bodyPatterns, matching(request.getBodyAsString()));
-
-		if (!matches) {
-			notifier().info(String.format("URL %s is match, but body is not: %s", request.getUrl(), request.getBodyAsString()));
-		}
-
-		return matches;
 	}
 
 	public String getUrlPattern() {
@@ -328,59 +230,5 @@ public class RequestPattern {
 		return Json.write(this);
 	}
 
-    private static final Function<Map.Entry<String,ValuePattern>,String> TO_KEYS_WHERE_VALUE_ABSENT = new Function<Map.Entry<String, ValuePattern>, String>() {
-        public String apply(Map.Entry<String, ValuePattern> input) {
-            return input.getValue().nullSafeIsAbsent() ? input.getKey() : null;
-        }
-    };
 
-    private static final Predicate<String> REMOVING_NULL = new Predicate<String>() {
-        public boolean apply(String input) {
-            return input != null;
-        }
-    };
-
-
-    private static Predicate<Map.Entry<String, ValuePattern>> matchHeadersIn(final Request request) {
-        return new Predicate<Map.Entry<String, ValuePattern>>() {
-            public boolean apply(Map.Entry<String, ValuePattern> headerPattern) {
-                ValuePattern headerValuePattern = headerPattern.getValue();
-                String key = headerPattern.getKey();
-                HttpHeader header = request.header(key);
-
-                boolean match = header.hasValueMatching(headerValuePattern);
-
-                if (!match) {
-                    notifier().info(String.format(
-                            "URL %s is match, but header %s is not. For a match, value should %s",
-                            request.getUrl(),
-                            key,
-                            headerValuePattern.toString()));
-                }
-
-                return match;
-            }
-        };
-    }
-
-    private Predicate<? super Map.Entry<String, ValuePattern>> matchQueryParametersIn(final Request request) {
-        return new Predicate<Map.Entry<String, ValuePattern>>() {
-            public boolean apply(Map.Entry<String, ValuePattern> entry) {
-                ValuePattern valuePattern = entry.getValue();
-                String key = entry.getKey();
-                Optional<QueryParameter> queryParam = Optional.fromNullable(request.queryParameter(key));
-                boolean match = queryParam.isPresent() && queryParam.get().hasValueMatching(valuePattern);
-
-                if (!match) {
-                    notifier().info(String.format(
-                            "URL %s is match, but query parameter %s is not. For a match, value should %s",
-                            request.getUrl(),
-                            key,
-                            valuePattern.toString()));
-                }
-
-                return match;
-            }
-        };
-    }
 }
